Initial commit
This is the initial implementation of the small tool converting zip to
tar files. Although there are many other tools like that, this one is
using only stdlib, is **not** meant to be a proper Python package, but
rather a single file script.

It also supports creating subdirs in the tar archive, which other tools
usually do not.

Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>